SERVER WIP

+++ server/routes/bookings.py
@@ -1,8 +1,10 @@
 #!/usr/bin/python3
 import sqlite3
 from flask import Blueprint, request, abort, make_response
-from modules import database, simple_jwt
-from routes.utils import logged_before_request, get_user_id
+from modules import simple_jwt
+from modules.database import get_db_conn
+from modules.utils import logged_before_request, get_role_perms, db_data_to_list
+
 
 route_bookings = Blueprint('route_bookings', __name__)
 route_bookings.before_request(logged_before_request)  # Check for login
@@ -14,23 +16,22 @@ route_bookings.before_request(logged_before_request)  # Check for login
 @route_bookings.route('/bookings', methods=['GET'])
 def booking_list():
     token_data = simple_jwt.read(request.headers.get('Authorization').split(' ')[1])
-    user_id = get_user_id(token_data, request.get_json())
-    with database.db_lock():
-        try:
+    user_perms = get_role_perms(token_data.get('role_id'))
+    if user_perms.get('booking_list', 0) or user_perms.get('booking_list_others', 0):
+        user_id = token_data.get('user')
+        if user_perms.get('booking_list_others', 0):
+            req_data = request.get_json()
+            if req_data:
+                user_id = req_data.get('user_id', token_data.get('user'))
+        with get_db_conn(True) as database:
             cursor = database.cursor()
             cursor.execute('SELECT id, user_id, lesson_id, status FROM bookings WHERE user_id = ?', [user_id])
             db_data = cursor.fetchall()
-            db_result = []
-            for row in db_data:
-                result_row = dict(map(lambda x, y: (x[0], y), cursor.description, row))
-                db_result.append(result_row)
-            result = make_response({'ok': True, 'data': db_result}, 200)
-        except sqlite3.Error as e:
-            result = make_response({'ok': False, 'error': str(e)}, 500)
-            print(e)
-        finally:
+            db_desc = cursor.description
             cursor.close()
-    return result
+        db_result = db_data_to_list(db_data, db_desc)
+        return make_response({'ok': True, 'data': db_result}, 200)
+    return abort(401)
 
 
 # --------------------------
@@ -39,25 +40,29 @@ def booking_list():
 @route_bookings.route('/bookings', methods=['POST'])
 def booking_add():
     token_data = simple_jwt.read(request.headers.get('Authorization').split(' ')[1])
+    user_perms = get_role_perms(token_data.get('role_id'))
     req_data = request.get_json()
     if req_data:
-        if token_data.get('role') != 3 or req_data.get('user_id') is None:
-            req_data.set('user_id', token_data.get('user'))
-        with database.db_lock():
-            try:
-                cursor = database.cursor()
-                cursor.execute('INSERT INTO bookings (user_id, lesson_id) VALUES (?, ?)',
-                               [req_data.get('user_id'), req_data.get('lesson_id')])
-                database.commit()
-                result = make_response({'ok': True}, 200)
-            except sqlite3.Error as e:
-                database.rollback()
-                result = make_response({'ok': False, 'error': str(e)}, 500)
-                print(e)
-            finally:
-                cursor.close()
-        return result
-    abort(400)
+        if user_perms.get('booking_add', 0) or user_perms.get('booking_add_other', 0):
+            user_id = token_data.get('user')
+            if user_perms.get('booking_add_other', 0) and req_data.get('user_id') is not None:
+                user_id = req_data.get('user_id')
+            with get_db_conn() as database:
+                try:
+                    cursor = database.cursor()
+                    cursor.execute('INSERT INTO bookings (user_id, lesson_id) VALUES (?, ?)',
+                                   [user_id, req_data.get('lesson_id')])
+                    last_id_inserted = cursor.lastrowid
+                    database.commit()
+                    result = make_response({'ok': True, 'data': last_id_inserted}, 200)
+                except sqlite3.Error as e:
+                    database.rollback()
+                    result = make_response({'ok': False, 'error': str(e)}, 500)
+                finally:
+                    cursor.close()
+            return result
+        return abort(401)
+    return abort(400)
 
 
 # --------------------------
@@ -66,18 +71,24 @@ def booking_add():
 @route_bookings.route('/bookings/<int:booking_id>', methods=['GET'])
 def booking_get(booking_id: int):
     token_data = simple_jwt.read(request.headers.get('Authorization').split(' ')[1])
-    user_id = get_user_id(token_data, request.get_json())
-    with database.db_lock():
-        cursor = database.cursor()
-        cursor.execute('SELECT user_id, lesson_id, status FROM bookings WHERE id = ? AND user_id = ?',
-                       [booking_id, user_id])
-        db_data = cursor.fetchone()
-        db_desc = cursor.description
-        cursor.close()
-    if db_data:
-        row = dict(map(lambda x, y: (x[0], y), db_desc, db_data))
-        return make_response({'ok': True, 'data': row}, 200)
-    return make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
+    user_perms = get_role_perms(token_data.get('role_id'))
+    if user_perms.get('booking_get', 0) or user_perms.get('booking_get_others', 0):
+        user_id = token_data.get('user')
+        if user_perms.get('booking_get_others', 0):
+            req_data = request.get_json()
+            if req_data:
+                user_id = req_data.get('user_id', token_data.get('user'))
+        with get_db_conn(True) as database:
+            cursor = database.cursor()
+            cursor.execute('SELECT * FROM bookings WHERE id = ? AND user_id = ?', [booking_id, user_id])
+            db_data = cursor.fetchone()
+            db_desc = cursor.description
+            cursor.close()
+        if db_data:
+            row = dict(map(lambda x, y: (x[0], y), db_desc, db_data))
+            return make_response({'ok': True, 'data': row}, 200)
+        return make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
+    return abort(401)
 
 
 # --------------------------
@@ -86,28 +97,32 @@ def booking_get(booking_id: int):
 @route_bookings.route('/bookings/<int:booking_id>', methods=['PUT'])
 def booking_update(booking_id: int):
     token_data = simple_jwt.read(request.headers.get('Authorization').split(' ')[1])
+    user_perms = get_role_perms(token_data.get('role_id'))
     req_data = request.get_json()
     if req_data:
-        req_data.remove('id')
-        user_id = get_user_id(token_data, req_data)
-        if token_data.get('role') != 3:
-            req_data.remove('user_id')
-            req_data.remove('lesson_id')
-        sql_str = 'UPDATE users SET ' + (', '.join("%s = ?" % v for v in req_data.keys())) \
-                  + ' WHERE id = ? AND user_id = ?'
-        with database.db_lock():
-            try:
-                cursor = database.cursor()
-                cursor.execute(sql_str, req_data.values() + [booking_id, user_id])
-                database.commit()
-                result = make_response({'ok': True}, 200)
-            except sqlite3.Error:
-                database.rollback()
-                result = make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
-            finally:
-                cursor.close()
-        return result
-    abort(400)
+        if user_perms.get('booking_update', 0) or user_perms.get('booking_update_others', 0):
+            req_data.remove('id')
+            user_id = req_data.get('user_id', token_data.get('user'))
+            if not user_perms.get('booking_update_others', 0):
+                user_id = token_data.get('user')
+                req_data.remove('user_id')
+                req_data.remove('lesson_id')
+            sql_str = 'UPDATE bookings SET ' + (', '.join("%s = ?" % v for v in req_data.keys())) \
+                      + ' WHERE id = ? AND user_id = ?'
+            with get_db_conn() as database:
+                try:
+                    cursor = database.cursor()
+                    cursor.execute(sql_str, req_data.values() + [booking_id, user_id])
+                    database.commit()
+                    result = make_response({'ok': True}, 200)
+                except sqlite3.Error:
+                    database.rollback()
+                    result = make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
+                finally:
+                    cursor.close()
+            return result
+        return abort(401)
+    return abort(400)
 
 
 # --------------------------
@@ -116,16 +131,23 @@ def booking_update(booking_id: int):
 @route_bookings.route('/bookings/<int:booking_id>', methods=['DELETE'])
 def booking_remove(booking_id: int):
     token_data = simple_jwt.read(request.headers.get('Authorization').split(' ')[1])
-    user_id = get_user_id(token_data, request.get_json())
-    with database.db_lock():
-        try:
-            cursor = database.cursor()
-            cursor.execute('DELETE FROM bookings WHERE id = ? AND user_id = ?', [booking_id, user_id])
-            database.commit()
-            result = make_response({'ok': True}, 200)
-        except sqlite3.Error:
-            database.rollback()
-            result = make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
-        finally:
-            cursor.close()
-    return result
+    user_perms = get_role_perms(token_data.get('role_id'))
+    if user_perms.get('booking_delete', 0) or user_perms.get('booking_delete_others', 0):
+        user_id = token_data.get('user')
+        if user_perms.get('booking_delete_others', 0):
+            req_data = request.get_json()
+            if req_data:
+                user_id = req_data.get('user_id', token_data.get('user'))
+        with get_db_conn() as database:
+            try:
+                cursor = database.cursor()
+                cursor.execute('DELETE FROM bookings WHERE id = ? AND user_id = ?', [booking_id, user_id])
+                database.commit()
+                result = make_response({'ok': True}, 200)
+            except sqlite3.Error:
+                database.rollback()
+                result = make_response({'ok': False, 'error': 'BOOKING_NOT_FOUND'}, 404)
+            finally:
+                cursor.close()
+        return result
+    return abort(401)
